% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lodes-package.R
\docType{package}
\name{lodes}
\alias{lodes}
\alias{lodes-package}
\title{Retrieve Data from LEHD Origin-Destination Employment Statistics Server}
\description{
Retrieve Data from LEHD Origin-Destination Employment Statistics Server
}
\author{
Bob Rudis (bob@rud.is)
}

