% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lodes.r
\name{read_lodes}
\alias{read_lodes}
\title{Retrieve a LODES file/data}
\usage{
read_lodes(state, type = c("origin", "residence", "workplace", "od", "rac",
  "wac"), segment, job_type, year, download_dir = getwd())
}
\arguments{
\item{state}{2-leter state code (auto-converted to lower case)}

\item{type}{one of "origin", "residence" or "workplace" or their short
equivalents "od", "rac", "wac"}

\item{segment}{segment identifier}

\item{job_type}{job identifier}

\item{year}{year to retrieve}

\item{download_dir}{location to cache the file (defaults to current directory)}
}
\value{
`tibble` of results if data file was found
}
\description{
Files are cached in \code{download_dir}
}
\note{
the full path to the downloaded file is displayed in message
}
\examples{
read_lodes("de", "od", "aux", "JT00", "2006")
read_lodes("ak", "residence", "S000", "JT00", "2002")
read_lodes("id", "workplace", "SE02", "JT03", "2013")
}
\references{
See \href{https://lehd.ces.census.gov/data/lodes/LODES7/LODESTechDoc7.2.pdf}{the codebook}
            for valid parameters.
}

