% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lodes.r
\name{read_xwalk}
\alias{read_xwalk}
\title{Retrieve geography crosswalk for a state}
\usage{
read_xwalk(state, download_dir = getwd())
}
\arguments{
\item{state}{state to read crosswalk for}

\item{download_dir}{location to cache the file (defaults to current directory)}
}
\description{
Reads the crosswalk file and returns a \code{tibble}.
Files are cached in \code{download_dir}
}
\examples{
read_xwalk("id")
}

